requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-openai
chromadb
pandas
tiktoken
python-dotenv
how